Refactored docs to support vscode doc viewer
