+++ src/index.ts
@@ -4,7 +4,9 @@ import { parseExp, parseProgram } from './parser.js'
 import { detailsToResult, ok, Result } from './result.js'
 import { scopeCheckExp, scopeCheckProgram } from './scope.js'
 
+export { Env, EnvEntry } from './lang.js'
 export * from './result.js'
+export { preludeEnv, internalLibs } from './runtime.js'
 export { expToString, stmtToString, progToString } from './lang.js'
 export * from './web.js'
