+++ src/lang.ts
@@ -13,14 +13,43 @@ const nlname = (value: string): Name => name(value, noRange())
 
 /* eslint-disable no-use-before-define */
 
-export type Entry = { value: Exp, source: string, range?: Range, documentation?: string }
-export const entry = (value: Exp, source: string, range?: Range, documentation?: string): Entry =>
-  ({ value, source, range, documentation })
+/**
+ * A `Doc` is a convenience class for constructing docstrings for library
+ * primitives.
+ */
+export class Doc {
+  /**
+   * 
+   * @param sig A docstring corresponding to the signature of the function.
+   * @param args An array of docstrings for each of the function's arguments.
+   * @param desc A prose description of the behavior of the function.
+   */
+  constructor (public sig: string, public args: string[], public desc: string) { }
+
+  /**
+   * @returns A string containing the docstring formatted in Markdown.
+   */
+  public docToMarkdown(): string {
+  return `
+~~~
+${this.sig.trim()}
+
+${this.args.map(arg => '  ' + arg.trim()).join('\n')}
+~~~
+
+${this.desc.trim()}
+  `.trim()
+  }
+}
+
+export type EnvEntry = { value: Exp, source: string, range?: Range, doc?: Doc }
+export const entry = (value: Exp, source: string, range?: Range, doc?: Doc): EnvEntry =>
+  ({ value, source, range, doc})
 
 export class Env {
-  entries: Map<string, Entry>
+  entries: Map<string, EnvEntry>
 
-  constructor (entries?: Iterable<[string, Entry]>) {
+  constructor (entries?: Iterable<[string, EnvEntry]>) {
     if (entries) {
       this.entries = new Map(entries)
     } else {
@@ -32,11 +61,11 @@ export class Env {
     return this.entries.has(key)
   }
 
-  public get (key: string): Entry | undefined {
+  public get (key: string): EnvEntry | undefined {
     return this.entries.get(key)
   }
 
-  public items (): Iterable<[string, Entry]> {
+  public items (): Iterable<[string, EnvEntry]> {
     return this.entries.entries()
   }
 
@@ -44,7 +73,7 @@ export class Env {
     return this.entries.keys()
   }
 
-  public append (key: string, value: Entry): Env {
+  public append (key: string, value: EnvEntry): Env {
     return new Env([...this.entries, [key, value]])
   }
